Finish audio system
The audio system, for the most part, is now finished. I'm still not entirely
happy with the code quality, especially with that of the StreamedSound class.
But I guess that comes with developing something like a game engine. I'm sure
I can make it better in the future, not to mention the sparse commits
containing large swaths of code. Those, on my part, are unforgivable. It is a
horrenouds habit of mine I have yet to break for projects like these usually.
Of course, I am more than sure this system has bugs in it, and that more will
be added or removed from it in the future. However, that will come in time.
What I have now I believe should be more than enough to suffice for beginning
base game development once I finish other components like engine config, game
config, graphics, etc.