Remove ML2 folder, use different dependency for Discord Rich Presence
I decided to use another dependency for Discord Rich Presence, as Maven
kept screaming at me that it was missing the Arikia implementation of
Discord RPC (even though it was already installed). So, I just went and
found an implementation that had an installation in another repo and
went on to use that. I may not even keep this included, unless it gets
a proper wrapper like the last version did. Although, I'm not really
sure the last version's wrapper was really "proper".