Add @BeResponsible
I decided to add this annotation to give me the peace of mind that I
intentionally left out checks in code for a reason. Usually this won't
be often, but it has to be done sometimes. Adding an annotation makes it
feel more "official", I guess? IDK, I'm probably just addicted to OOP.