Add provider package and classes
I decided to add these as a way to guarantee a more dynamic approach for
different parts of the code that require some sort of matrix when
rendering. This allows a camera to be used for a view matrix, but
prevents a camera from being the ONLY thing that can be used for a view
matrix. This will either come in handy or prove to be over-engineering
OOP paranoia in the future.