Add optional() to Uniform.java
I realized that not all programs will need some shaders that could be
considered global (for example, the orthographic matrix or view matrix).
After discovering the OpenGL allows -1 to be given as a value for
glUniform, I implemented the optional() attribute for this annotation.
If set to true, any missing uniform will have their @Uniform annotated
field set to -1.