Update ArdenusManifest.java
This update brings this class up to my current standards of code for my
projects, and also got rid of some methods that I thought were out of
date and should go somewhere else. For example, storing the title in the
game manifest I think is not a good idea. What if the game's name is
different in a different language?