Create Analog class for input system
The purpose of this class will be to provide an interface for providing
analog input values. That being, values that are not just simply PRESSED,
NOT PRESSED, HELD DOWN, etc. This commit does not fully complete the
functionality for the Analog system, but it introduces most of the base
classes needed for it. The next commits will introduce adapter support
for them, so that they can be used with input devices.