Reduce template usage in DeviceAdapter
The amount of templates that were present in the DeviceAdapter class was
becoming an ugly sight. Not to mention, it did not bring the functionality
needed for the class. In later commits, I plan to fix the shoddy workaround
of casting everything to the desire type via the use of a ButtonAdapter and
AnalogAdapter annotation.