Add support for Sony PlayStation 4 controllers via GLFW
While this commit adds PS4 controller support, there is one glaring issue that
I have found before making this commit. That is, when moving the left analog
stick to the left on its X-axis, the Y-axis of that analog stick reports that
it is being moved upwards (or downwards) by over 86% on a 0.0-1.0 scale. This
issue occurred with both controllers that came with my PS4 when I got it back
in 2013. After doing some online research, I found that others seemed to be
having this issue. I will do some research to see if my implementation needs
to be updated, or if both of my controllers are just borked. Other than this
single issue with the left analog stick (the right analog stick does not have
this problem, at least nowhere near as badly), everything is functional.