Overhaul some code for college demo
I don't really like this commit (or a few commits I'm about to make, God
in heaven forgive me), but these will have to do as I have a
presentation in but hours from this commit. By this point, I just need
it to *work*. Is this what it's like to be a game developer in
production?