Fix erroneous use of getDeclaredFields() in InputDevice, DeviceAdapter
This erroneous use of the method made it so that input devices would not
inherit the device features of the class that they extended. As such, I have
decided to make it an error to declare device feature marked with the
FeaturePresent annotation as anything other than public. However, I will
implement the proper checks for this in a later commit.