Add missing XBOX controller GUID
This GUID added is that of my Xbox Series X controller. It seems as though the
SDL controller database does not have this one? I used the SDL mapping tool,
and it recognized my XBOX controller as it should. Maybe I should create an
issue for it sometime? I hope I'm not using these GUIDs incorrectly!